+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- change `takeDomCapture` to capture dom only of the provided context
 
 ## 12.9.3 - 2020/12/15